ToDoList/requirements.txt
Django>=4.2
djangorestframework
gunicorn
pymysql
cryptography==44.0.0
drf-yasg
django-cors-headers